DEV-5909: Relocate integration card components
