+++ spa/src/components/dashboard/MailchimpConnectionStatus.test.tsx
@@ -8,7 +8,7 @@ import useUser from 'hooks/useUser';
 jest.mock('hooks/usePreviousState');
 jest.mock('hooks/useUser');
 jest.mock('components/common/Modal/AudienceListModal/AudienceListModal');
-jest.mock('components/common/IntegrationCard/MailchimpIntegrationCard/MailchimpModal/MailchimpModal');
+jest.mock('components/settings/integration/IntegrationCard/MailchimpIntegrationCard/MailchimpModal/MailchimpModal');
 jest.mock('hooks/useConnectMailchimp');
 jest.mock('components/common/GlobalLoading/GlobalLoading');
 jest.mock('notistack', () => ({
