+++ spa/src/components/settings/Integration/IntegrationCard/IntegrationCardHeader/__mocks__/IntegrationCardHeader.tsx
