+++ spa/src/components/settings/Integration/IntegrationCard/MailchimpIntegrationCard/MailchimpIntegrationCard.test.tsx
