+++ spa/src/components/settings/Integration/IntegrationCard/ModalUpgradePrompt/ModalUpgradePrompt.styled.ts
