+++ spa/src/components/settings/Integration/IntegrationCard/ModalUpgradePrompt/__mocks__/ModalUpgradePrompt.tsx
