+++ spa/src/components/settings/Integration/IntegrationCard/SlackIntegrationCard/index.ts
