+++ spa/src/components/settings/Integration/IntegrationCard/StripeIntegrationCard/StripeIntegrationCard.tsx
