+++ spa/src/components/settings/Integration/IntegrationCard/StripeIntegrationCard/index.ts
