+++ spa/src/components/settings/Integration/IntegrationCard/__mocks__/IntegrationCard.tsx
