+++ spa/src/components/settings/Integration/IntegrationCard/index.ts
